% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic.R
\name{all_endemic}
\alias{all_endemic}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
all_endemic(sim)
}
\arguments{
\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_cr}()},
\code{\link[DAISIE]{DAISIE_sim_time_dep}()},
\code{\link[DAISIE]{DAISIE_sim_cr_shift}()},
\code{\link[DAISIE]{DAISIE_sim_trait_dep}()} or other input simulation
in the DAISIE sim format.}
}
\value{
Boolean
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
