% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic_clades.R
\name{all_endemic_clades}
\alias{all_endemic_clades}
\title{Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.}
\usage{
all_endemic_clades(island)
}
\value{
Boolean
}
\description{
Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.
}
\author{
Joshua W. Lambert
}
