% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_recols.R
\name{any_recols}
\alias{any_recols}
\title{Determines if the island contains any recolonisations that survived to the
present}
\usage{
any_recols(sim)
}
\arguments{
\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_cr}()},
\code{\link[DAISIE]{DAISIE_sim_time_dep}()},
\code{\link[DAISIE]{DAISIE_sim_cr_shift}()},
\code{\link[DAISIE]{DAISIE_sim_trait_dep}()} or other input simulation
in the DAISIE sim format.}
}
\value{
Boolean
}
\description{
Determines if the island contains any recolonisations that survived to the
present
}
