% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_factor_key.R
\name{append_factor_key}
\alias{append_factor_key}
\title{Create key to partition results by scenario}
\usage{
append_factor_key(scenario_res, partition_by)
}
\arguments{
\item{scenario_res}{Data frame as returned by
  \code{\link{load_param_space}()}which is identical to a scenarios'
  parameter space as as read by \code{\link{calc_ed95_for_plots}()}. In
  addition to the columns in the parameter spaces, 5 columns are appended
  with the ed95 statistic obtained for each set parameter set:
\itemize{
  \item{\code{$ed95_spec_nltt}}
  \item{\code{$ed95_endemic_nltt}}
  \item{\code{$ed95_nonendemic_nltt}}
  \item{\code{$ed95_num_spec}}
  \item{\code{$ed95_num_col}}
}}

\item{partition_by}{String with the feature by which the results should be
  grouped for plotting. Options are:
\itemize{
  \item{\code{"hyperparameters"}}
  \item{\code{"gradient"}}
  \item{\code{"sample_parameters"}}
  \item{\code{"land_bridge_multiplier"}}
}}
}
\value{
A data frame with a parameter space and corresponding ed95 results
  returned by \code{\link{calc_ed95_for_plots}()}. In addition, a column
  \code{$key} is appended at the right hand side with the factors by which
  the data should be grouped for plotting.
}
\description{
Create key to partition results by scenario
}
\author{
Pedro Santos Neves
}
\keyword{internal}
