% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed95_for_plots.R
\name{calc_ed95_for_plots}
\alias{calc_ed95_for_plots}
\title{Calculate ED95 statistic for all parameter sets of a parameter space}
\usage{
calc_ed95_for_plots(folder_path, scenario)
}
\arguments{
\item{folder_path}{A file path (use \code{file.path()} for formatting
convenience) where all the parameter sets results of a given parameter
space are stored in .rds format.}
}
\value{
A data frame which is identical to each scenarios' parameter spaces
  as as read by \code{\link{load_param_space}()}. In addition to the columns
  in the parameter spaces, 5 columns are appended with the ed95 statistic
  obtained for each set parameter set.
\itemize{
  \item{\code{$ed95_spec_nltt}}
  \item{\code{$ed95_endemic_nltt}}
  \item{\code{$ed95_nonendemic_nltt}}
  \item{\code{$ed95_num_spec}}
  \item{\code{$ed95_num_col}}
}
}
\description{
This function will load the results of each parameter set from a given
parameter space and compute all associated ED95 statistics. All parameter
sets of the parameter space must be in the same folder, located in
\code{folder_path}. The return type is appropriate for the main plot
functions.
}
\author{
Pedro Neves, Joshua W. Lambert
}
