% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed95_param_set.R
\name{calc_ed95_param_set}
\alias{calc_ed95_param_set}
\title{Calculate ED95 statistic for all results in folder}
\usage{
calc_ed95_param_set(folder_path)
}
\arguments{
\item{folder_path}{A file path (use \code{file.path()} for formatting
convenience) where all the parameter sets results of a given parameter
space are stored in .RData format.}
}
\value{
A data frame with ED95 statistics containing 7 columns:
\itemize{
  \item{\code{$param_space_name}}
  \item{\code{$param_set}}
  \item{\code{$ed95_spec_nltt}}
  \item{\code{$ed95_endemic_nltt}}
  \item{\code{$ed95_nonendemic_nltt}}
  \item{\code{$ed95_num_spec}}
  \item{\code{$ed95_num_col}}
}
}
\description{
This function will load the results of each parameter set from all
parameter spaces and compute all associated ED95 statistics. All parameter
sets must be inside named sub folders inside \code{folder_path}. Return type
is appropriate for functions dealing with run time vs ED95 correlations.
}
\author{
Pedro Neves, Joshua W. Lambert
}
\keyword{internal}
