% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.R
\name{calc_error}
\alias{calc_error}
\title{Calculates error metrics between two simulations}
\usage{
calc_error(sim_1, sim_2, sim_pars, replicates, distance_method)
}
\arguments{
\item{sim_1}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()}.}

\item{sim_2}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()} or
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()}.}

\item{sim_pars}{A list of simulation parameters.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{distance_method}{From the nLTT package.
How the difference between the two nLTTs is summed:
\itemize{
 \item{\code{"abs"}: }{the absolute distance between the two nLTTs is
   summed}
 \item{\code{"squ"}: }{the squared distance between the two nLTTs is summed}
}}
}
\value{
A list with five error metrics:
\describe{
  \item{\code{$spec_nltt_error}}{Numeric vector with the nltt error between
    the two simulations.}
  \item{\code{$num_spec_error}}{Numeric vector with the difference in the
    number of species at the end of the simulation between the two
    simulations.}
  \item{\code{$num_col_error}}{Numeric vector with the difference in the
    number of colonists at the end of the simulation between the two
    simulations.}
  \item{\code{$endemic_nltt_error}}{Numeric vector with the nltt error of the
    endemic species between the two simulations.}
  \item{\code{$nonendemic_nltt_error}}{Numeric vector with the nltt error of
    the non-endemic species between the two simulations.}
}
}
\description{
Calculates error in number of species and colonists, number of
species-through-time, number of endemics-through-time and
non-endemics-through-time.
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
