% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_js_dist.R
\name{calc_js_dist}
\alias{calc_js_dist}
\title{Calculates Jensen-Shannon distance between two distributions}
\usage{
calc_js_dist(
  spec_error,
  endemic_error,
  nonendemic_error,
  spec_baseline_error,
  endemic_baseline_error,
  nonendemic_baseline_error
)
}
\arguments{
\item{spec_error}{A list with three elements each with a numeric
vector.}

\item{endemic_error}{A list with one element with a numeric vector.}

\item{nonendemic_error}{A list with one element with a numeric vector.}

\item{spec_baseline_error}{A list with three elements each with
a numeric vector.}

\item{endemic_baseline_error}{A list with one element with a numeric
vector.}

\item{nonendemic_baseline_error}{A list with one element with a numeric
vector.}
}
\value{
a list of numerics
}
\description{
Calculates Jensen-Shannon distance between two distributions
}
\author{
Pedro Neves, Joshua Lambert
}
