% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_js_dist.R
\name{calc_js_div}
\alias{calc_js_div}
\title{Calculates the Jensen-Shannon divergence between two vectors}
\usage{
calc_js_div(p, q)
}
\arguments{
\item{p}{Numeric vector to calculate divergence metric.}

\item{q}{Numeric vector to calculate divergence metric.}
}
\value{
numeric
}
\description{
Calculates the Jensen-Shannon divergence between two vectors
}
\references{
Lamberti, P. W., et al. "Metric character of the quantum
  Jensen-Shannon divergence." Physical Review A 77.5 (2008): 052311.
}
\author{
Pedro Neves, Joshua Lambert
}
