% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_metrics.R
\name{calc_ks_distance}
\alias{calc_ks_distance}
\title{Calculate the Kolmogorov-Smirnov distance between two distributions}
\usage{
calc_ks_distance(vec_1, vec_2)
}
\arguments{
\item{vec_1}{A numeric vector. Usually a distribution of nLTT statistics
or similar vectors as obtained within
\code{\link{calc_error_metrics}()}. Distance is calculated with
\code{vec_2}.}

\item{vec_2}{A numeric vector. Usually a distribution of nLTT statistics
or similar vectors as obtained within
\code{\link{calc_error_metrics}()}. Distance is calculated with
\code{vec_1}.}
}
\value{
A numeric, with the Kolmogorov-Smirnov distance of the two
 distributions.
}
\description{
Calculate the Kolmogorov-Smirnov distance between two distributions
}
\references{
Kolmogorov, A. (1933) Sulla determinazione empirica di una
legge di distribuzionc, 1st. Ital. Attuari. G.. 4, 1–11.

Razali and Wah (2011) "Power comparisons of shapiro-wilk,
kolmogorov-smirnov, lilliefors and anderson-darling tests"
Journal of Statistical Modeling and Analytics 2, 21-33.
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
