% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_spec.R
\name{calc_max_spec}
\alias{calc_max_spec}
\title{Obtain the number of species in the largest clade of each island}
\usage{
calc_max_spec(novel_sim)
}
\arguments{
\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}
}
\value{
A numeric vector with the number of species in the largest clade
 in a given island, per replicate.
}
\description{
Obtain the number of species in the largest clade of each island
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
