% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_run_success_ratio.R
\name{calc_run_success_ratio}
\alias{calc_run_success_ratio}
\title{Calculate success rate of \code{\link{run_robustness}()} HPCC runs}
\usage{
calc_run_success_ratio(logs_folder_path)
}
\arguments{
\item{logs_folder_path}{A string with the path for a directory containing the
Peregrine HPCC log files.}
}
\value{
A data frame containing 4 columns and as many rows as the parameter
  spaces found on the log files. Columns are as follows:
  \itemize{
    \item{\code{"param_space"}: }{string with the name of the parameter
    spaces}
    \item{\code{"success_rate"}: }{A numeric between 0 and 1 with the ratio
    of successful runs of each parameter space}
    \item{\code{"failed_rate"}: }{A numeric between 0 and 1 with the ratio
    of failed runs of each parameter space}
    \item{\code{"timeout_rate"}: }{A numeric between 0 and 1 with the ratio
    of timed out runs of each parameter space}
 }
}
\description{
Calculate success rate of \code{\link{run_robustness}()} HPCC runs
}
\author{
Pedro Santos Neves
}
