% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime_ed95_cor.R
\name{calc_runtime_ed95_cor}
\alias{calc_runtime_ed95_cor}
\title{Calculate correlation between ED95 statistics and runtime}
\usage{
calc_runtime_ed95_cor(runtime_params, ed95_param_sets)
}
\arguments{
\item{runtime_params}{A data frame with runtime of each parameter set in each
parameter space as returned by \code{\link{get_runtime_params}()}.}

\item{ed95_param_sets}{A data frame with all ED95 statistics of each
parameter set in each parameter space as returned by
\code{\link{calc_ed95_param_set}()}.}
}
\value{
A data frame with the Pearson correlation coefficient between runtime
  and each of the ED95 metrics.
}
\description{
Calculate correlation between ED95 statistics and runtime
}
\author{
Pedro Santos Neves
}
