% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_runtime_log.R
\name{calc_runtime_log}
\alias{calc_runtime_log}
\title{Calculate job elapsed time from log file}
\usage{
calc_runtime_log(log_file_path)
}
\arguments{
\item{log_file_path}{A string with the path for a standard Peregrine HPCC log
file.}
}
\value{
An atomic numeric vector with the elapsed run time of the associated
  job.
}
\description{
Obtains the elapsed job time in seconds from a standard Peregrine HPCC log
file
}
\examples{
\dontrun{
runtime <- calc_runtime_log(log_file_path = "logs/robustness-20596511.log")
}
}
\author{
Pedro Neves
}
