% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime_param_corr.R
\name{calc_runtime_param_corr}
\alias{calc_runtime_param_corr}
\title{Calculate correlation between parameters}
\usage{
calc_runtime_param_corr(logs_folder_path)
}
\arguments{
\item{logs_folder_path}{}
}
\value{

}
\description{
Calculate correlation between parameters
}
