% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stat_diff.R
\name{calc_stat_diff}
\alias{calc_stat_diff}
\title{Calculate stat diff}
\usage{
calc_stat_diff(folder_path, param_set_range = NULL)
}
\value{
Numeric vector with computed statistic.
}
\description{
Calculate stat diff
}
