% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_area_plot_input}
\alias{check_area_plot_input}
\title{Check if parameters for area plotting are valid}
\usage{
check_area_plot_input(
  totaltime_1,
  totaltime_2,
  area_pars_1,
  area_pars_2,
  island_ontogeny_1,
  island_ontogeny_2,
  sea_level_1,
  sea_level_2,
  resolution,
  overlay_sea_level
)
}
\arguments{
\item{totaltime_1}{Numeric with time for which island 1 would be simulated.
Usually this will correspond to a time that's smaller than the total
island age.}

\item{totaltime_2}{Time for which island 2 would be simulated. Usually this
will correspond to a time that's smaller than the total island age.
Defaults to \code{NULL}, in which case this island is not plotted.}

\item{area_pars_1}{List of area and sea-level parameters for island 1 as
created by \code{\link[DAISIE]{create_area_pars}}.}

\item{area_pars_2}{List of area and sea-level parameters for island 1 as
created by \code{\link[DAISIE]{create_area_pars}}. Defaults to \code{NULL},
in which case the second area is not plotted.}

\item{island_ontogeny_1}{A string describing the type of
island ontogeny of island 1.
Can be \code{"const"}, \code{"beta"} for a beta function describing area
through time.}

\item{island_ontogeny_2}{A string describing the type of
island ontogeny of island 2.
Can be \code{"const"}, \code{"beta"} for a beta function describing area
through time. Defaults to \code{NULL}, in which case the second island is
not plotted.}

\item{sea_level_1}{A string describing the type of sea level for island 1.
Can be \code{"const"} or \code{"sine"} for a sine function describing area
through time.}

\item{sea_level_2}{A string describing the type of sea level for island 2.
Can be \code{"const"} or \code{"sine"} for a sine function describing area
through time. Defaults to \code{NULL}, in which case the second island is
not plotted.}

\item{resolution}{A float between 0 and 1 which controls the resolution of
the island area calculated for plotting. Defaults to \code{0.001} and is
bound at \code{0.00001} to prevent unwieldy slowdowns.}

\item{overlay_sea_level}{Logical, controling if sea level curves
should be overlaid on top of corresponding ontogeny curves.}
}
\value{
Nothing, breaks with helpful error message if input is wrong.
}
\description{
Asserts all individual arguments for plotting are consistent. If any isn't,
a helpful error message is returned.
}
\seealso{
DAISIE_plot_areas
}
\author{
Pedro Neves, Joshua Lambert
}
