% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_metrics.R
\name{check_calc_error_metrics_input}
\alias{check_calc_error_metrics_input}
\title{Checks calc_error_metrics_input is correct}
\usage{
check_calc_error_metrics_input(
  spec_nltt_error,
  num_spec_error,
  num_col_error,
  endemic_nltt_error,
  nonendemic_nltt_error,
  spec_baseline_nltt_error,
  num_spec_baseline_error,
  num_col_baseline_error,
  endemic_baseline_nltt_error,
  nonendemic_baseline_nltt_error
)
}
\value{
Nothing. Throws error if input is not correct
}
\description{
Checks calc_error_metrics_input is correct
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
