% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{check_create_results_folder}
\alias{check_create_results_folder}
\title{Checks and create results folder for output}
\usage{
check_create_results_folder(param_space_name, save_output)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}, \code{"trait"}.}

\item{save_output}{A boolean to determine whether to save or return output.}
}
\value{
Helpful messages with info on folder status. Tries to create
  folder if needed.
}
\description{
Checks and create results folder for output
}
\seealso{
Other I/O: 
\code{\link{create_output_file_name}()},
\code{\link{load_novel_section}()},
\code{\link{save_output}()}
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\concept{I/O}
\keyword{internal}
