% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_constraints.R
\name{check_ml_tolerance}
\alias{check_ml_tolerance}
\title{Checks if MLE results converge}
\usage{
check_ml_tolerance(ml_res_initpars_1, ml_res_initpars_2)
}
\arguments{
\item{ml_res_initpars_1}{Numeric data frame.
Results from MLE run of geodynamics with the first set of initial
parameters.}

\item{ml_res_initpars_2}{Numeric data frame.
Results from MLE run of geodynamics with the second set of initial
parameters.}
}
\value{
Logical \code{TRUE} if criteria are met, \code{FALSE} if not.
}
\description{
Checks if MLE results converge
}
