% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param_space_time_dependent.R
\name{check_param_space_time_dependent}
\alias{check_param_space_time_dependent}
\title{Checks conditioning and diversity-dependent model for time-dependent
simulations}
\usage{
check_param_space_time_dependent(param_space_name, param_set, replicates = 100)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"}, or \code{"nonoceanic_sea_level"}}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{number of replicates for the initial simulation.}
}
\description{
Checks conditioning and diversity-dependent model for time-dependent
simulations
}
