% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{continental_land_bridge_iw}
\alias{continental_land_bridge_iw}
\title{Parameter space of the continental land-bridge island-wide
diversity-dependent analysis}
\format{
A data frame with 128 rows and 16 variables:
\describe{
  \item{time}{Numeric defining the length of the simulation in time units}
  \item{M}{Numeric defining the size of mainland pool, i.e. the number of
  species that can potentially colonize the island}
  \item{lac_1}{lambda^c (cladogenesis rate)}
  \item{mu_1}{mu (extinction rate)}
  \item{K_1}{K (carrying capacity), set K=Inf for diversity independence}
  \item{gam_1}{gamma (immigration rate)}
  \item{laa_1}{lambda^a (anagenesis rate)}
  \item{x_s}{the probability of sampling a species from the mainland}
  \item{x_nonend}{the probability of the species sampled from the mainland
  being nonendemic}
  \item{shift_times}{Numeric vector specifying when the rate shifts occur
  before the present}
  \item{divdepmodel}{Option divdepmodel = 'CS' runs a model with
  clade-specific carrying capacity, where diversity-dependence operates only
  within single clades, i.e. only among species originating from the same
  mainland colonist. Option divdepmodel = 'IW' runs a model with island-wide
  carrying capacity, where diversity-dependence operates within and among
  clades.}
  \item{lac_2}{lambda^c (cladogenesis rate) for rate set 2 in rate shift
  model}
  \item{mu_2}{mu (extinction rate) for rate set 2 in rate shift model}
  \item{K_2}{K (K (carrying capacity) for rate set 2 in rate shift model, set
  K=Inf for diversity independence}
  \item{gam_2}{gamma (immigration rate) for rate set 2 in rate shift model}
  \item{laa_2}{ambda^a (anagenesis rate) for rate set 2 in rate shift model}
}
}
\description{
A dataset containing the DAISIE parameters for 128 parameter
sets.
}
