% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{create_output_file_name}
\alias{create_output_file_name}
\title{Generated name for saving output of \code{run_robustness()}}
\usage{
create_output_file_name(
  param_space_name,
  param_set,
  sim_constraints,
  ml_constraints
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{sim_constraints}{Boolean. \code{TRUE} if simulation constraints pass,
\code{FALSE} otherwise.}

\item{ml_constraints}{Boolean. \code{TRUE} if MLE constraints pass,
\code{FALSE} otherwise.}
}
\value{
Character with name indicating the name of the parameter space,
  the numeric id of the param_set and whether the run passed constraints.
}
\description{
Generated name for saving output of \code{run_robustness()}
}
\examples{
testit::assert(
  DAISIErobustness:::create_output_file_name(
    param_space_name = "oceanic_ontogeny",
    param_set = 1,
    sim_constraints = TRUE,
    ml_constraints = TRUE
  ) == "passed_cond_oceanic_ontogeny_param_set_1.RData"
)
}
\seealso{
Other I/O: 
\code{\link{check_create_results_folder}()},
\code{\link{load_novel_section}()},
\code{\link{save_output}()}
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\concept{I/O}
\keyword{internal}
