% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIErobustness-package.R
\docType{package}
\name{DAISIErobustness-package}
\alias{DAISIErobustness}
\alias{DAISIErobustness-package}
\title{DAISIErobustness: Test the robustness of DAISIE to geodynamics}
\description{
Conducts a robustness test of the island biogeography DAISIE model
    to the effects of geodynamics, such as island ontogeny, sea-level changes
    and land-bridge dynamics. Contains accessory code to run at the Peregrine 
    HPCC at the University of Groningen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Neves-P/DAISIErobustness}
  \item Report bugs at \url{https://github.com/Neves-P/DAISIErobustness/issues}
}

}
\author{
\strong{Maintainer}: Joshua Lambert \email{j.w.l.lambert@rug.nl}

Authors:
\itemize{
  \item Pedro Neves \email{p.m.santos.neves@rug.nl}
}

}
\keyword{internal}
