% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_constraints.R
\name{decide_best_pars}
\alias{decide_best_pars}
\title{Check if MLE results converge and chooses best MLE output}
\usage{
decide_best_pars(ml_res_initpars_1, ml_res_initpars_2)
}
\arguments{
\item{ml_res_initpars_1}{Numeric data frame.
Results from MLE run of geodynamics with the first set of initial
parameters.}

\item{ml_res_initpars_2}{Numeric data frame.
Results from MLE run of geodynamics with the second set of initial
parameters.}
}
\value{
List with item per replicate containing:
  \itemize{
    \item{\code{pars_to_use}: Single row data frame with the parameter
    set from the MLE with the highest obtained loglik value.}
    \item{\code{loglik_tolerance_check}: Character
    \code{"fail loglik difference"} which indicates that the absolute
    difference between the two compared logliks is greater than the tolerance
    value (1e-6) or \code{"pass loglik difference"} when the absolute
    difference is smaller than the tolerance.}
    \item{\code{absolute_loglik_difference}: Numeric with the absolute
    difference between the two logliks.}
    \item{\code{pars_tolerance_check}: Character
    \code{"fail pars difference"} if the absolute difference between any of
    the the two compared estimated parameter sets is greater than the
    tolerance value (1e-6) or \code{"pass pars difference"} when the absolute
    difference is smaller than the tolerance.}
    \item{\code{absolute_pars_difference}: List of numeric vectors with the
    absolute difference calculated between the two parameter sets.}
  }
}
\description{
Since two MLE runs with different initial optimizer parameters
are done, the best parameter set between the two must be decided. This is
done by choosing the set for which the highest loglikelihood value was
obtained.
}
\details{
To keep track of the differences between the two estimated sets,
it is also determined if the estimated parameters differ by a value greater
than the normally used tolerance value for the ML optimization (1e-6). The
same is also done regarding the absolute differences in loglikelihoods.
The flow of \code{\link{run_robustness}()} is not affected by the
comparisons with a tolerance, rather these values are kept for downstream
QC of results.
}
\seealso{
\code{\link[DAISIE]{DAISIE_ML}()}, \code{\link{calc_ml}()}
}
\author{
Joshua Lambert, Pedro Neves
}
