% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  param_space_name,
  param_set,
  replicates,
  param_space,
  simulation_pars,
  simulations,
  ml,
  simulations_1,
  simulations_2,
  verbose,
  path
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}

\item{param_space}{tibble of the parameter space.}

\item{simulation_pars}{A list of simulation parameters.}

\item{simulations}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{ml}{Output from \code{\link{DAISIE_ML_CS}}}

\item{simulations_1}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{simulations_2}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{verbose}{logical determining whether to print
simulation statements to the console.}

\item{path}{String with path where log files are located.}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
