% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param_set.R
\name{extract_param_set}
\alias{extract_param_set}
\title{Extracts a set of parameters from the parameter space for the
simulation}
\usage{
extract_param_set(param_space_name, param_space, param_set)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. See
\code{\link{is_param_space_name}()} for options.}

\item{param_space}{A data frame of the parameter space.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}
}
\value{
a list of simulation parameters
}
\description{
Extracts a set of parameters from the parameter space for the
simulation
}
\author{
Joshua Lambert, Pedro Santos Neves, Shu Xie
}
