% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodynamic_sim.R
\name{geodynamic_sim}
\alias{geodynamic_sim}
\title{Runs DAISIE simulation with geodynamics}
\usage{
geodynamic_sim(param_space_name, sim_pars, replicates)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{sim_pars}{A list of simulation parameters.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}
}
\value{
A list output from \code{\link[DAISIE]{DAISIE_sim_constant_rate}},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}}, or
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}}
}
\description{
Runs DAISIE simulation with geodynamics
}
\author{
Joshua Lambert, Pedro Neves
}
