% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodynamic_simulations.R
\name{geodynamic_simulations}
\alias{geodynamic_simulations}
\title{Runs DAISIE simulation with geodynamics}
\usage{
geodynamic_simulations(param_space_name, simulation_pars, replicates, verbose)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{simulation_pars}{A list of simulation parameters.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}

\item{verbose}{logical determining whether to print
simulation statements to the console.}
}
\value{
A list output from \code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}}, or
\code{\link{DAISIE_sim_constant_rate_shift}}
}
\description{
Runs DAISIE simulation with geodynamics
}
