% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_results}
\alias{get_available_results}
\title{Get param set IDs stored in results folder}
\usage{
get_available_results(param_space_name)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}
}
\value{
A vector with param_set IDs
}
\description{
Get param set IDs stored in results folder
}
\author{
Joshua Lambert, Pedro Neves
}
