% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_submitted_param_sets}
\alias{get_submitted_param_sets}
\title{Get param_space number from log file}
\usage{
get_submitted_param_sets(path)
}
\arguments{
\item{path}{String with path where log files are located.}
}
\value{
Numeric vector with param_space of jobs in the log files
}
\description{
Get param_space number from log file
}
