% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integration_error.R
\name{is_integration_error_log}
\alias{is_integration_error_log}
\title{Checks if log file shows integration error}
\usage{
is_integration_error_log(log_lines)
}
\arguments{
\item{log_lines}{Character vector containing the contents of the log files
as read by \code{\link{read_log_file}()}.}
}
\value{
A boolean, \code{TRUE} if integration error occured, \code{FALSE} if
not.
}
\description{
Checks if log file shows integration error
}
\author{
Pedro Santos Neves
}
