% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_novel_sim_outputs}
\alias{is_novel_sim_outputs}
\title{Measures if the input is a valid collection of simulation
outputs.}
\usage{
is_novel_sim_outputs(novel_sim)
}
\arguments{
\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}
}
\value{
TRUE if the input is a valid collection of simulation
outputs.
}
\description{
Measures if the input is a valid collection of simulation
outputs.
}
\details{
Adapted from \code{\link[DAISIE]{is_simulation_outputs}()}
}
\author{
Richèl J.C Bilderbeek, Pedro Santos Neves
}
