% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_param_space_name.R
\name{is_param_space_name}
\alias{is_param_space_name}
\title{Checks whether the parameter space name give is one of the parameter spaces
that are included in the DAISIErobustness package}
\usage{
is_param_space_name(param_space_name)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. See
\code{\link{is_param_space_name}()} for options.}
}
\value{
Boolean
}
\description{
Checks whether the parameter space name give is one of the parameter spaces
that are included in the DAISIErobustness package
}
\examples{
bool <- is_param_space_name("oceanic_ontogeny_cs")
}
