% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{load_novel_section}
\alias{load_novel_section}
\title{Load intermediate novel sim results to continue pipeline}
\usage{
load_novel_section(param_space_name, param_set)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. See
\code{\link{is_param_space_name}()} for options.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}
}
\value{
List with output from \code{\link{run_novel_sim}()}.
}
\description{
Load intermediate novel sim results to continue pipeline
}
\seealso{
Other I/O: 
\code{\link{check_create_folders}()},
\code{\link{create_output_file_name}()},
\code{\link{save_output}()}
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\concept{I/O}
