% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ml.R
\name{mean_geodynamic_ml}
\alias{mean_geodynamic_ml}
\title{Calculate mean of ML estimates}
\usage{
mean_geodynamic_ml(ml_res_initpars_1, ml_res_initpars_2)
}
\arguments{
\item{ml_res_initpars_1}{Numeric data frame.
Results from MLE run of geodynamics with the first set of initial
parameters.}

\item{ml_res_initpars_2}{Numeric data frame.
Results from MLE run of geodynamics with the second set of initial
parameters.}
}
\value{
Numeric list with elementwise means of ML runs with different initial
  parameters.
}
\description{
Calculate mean of ML estimates
}
