% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonoceanic_di}
\alias{nonoceanic_di}
\title{Parameter space of the nonoceanic diversity-independent analysis}
\format{
A data frame with 192 rows and 10 variables:
\describe{
  \item{time}{Numeric defining the length of the simulation in time units}
  \item{M}{Numeric defining the size of mainland pool, i.e. the number of
  species that can potentially colonize the island}
  \item{lac}{lambda^c (cladogenesis rate)}
  \item{mu}{mu (extinction rate)}
  \item{K}{K (carrying capacity), set K=Inf for diversity independence}
  \item{gam}{gamma (immigration rate)}
  \item{laa}{lambda^a (anagenesis rate)}
  \item{x_s}{the probability of sampling a species from the mainland}
  \item{x_nonend}{the probability of the species sampled from the mainland
  being nonendemic}
  \item{divdepmodel}{Option divdepmodel = 'CS' runs a model with
  clade-specific carrying capacity, where diversity-dependence operates only
  within single clades, i.e. only among species originating from the same
  mainland colonist. Option divdepmodel = 'IW' runs a model with island-wide
  carrying capacity, where diversity-dependence operates within and among
  clades.}
}
}
\usage{
nonoceanic_di
}
\description{
A dataset containing the DAISIE parameters for 192 parameter sets.
}
\keyword{datasets}
