% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novel_sim.R
\name{novel_sim}
\alias{novel_sim}
\title{Runs DAISIE simulation with a novel model}
\usage{
novel_sim(param_space_name, sim_pars, replicates)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait_CES"},
\code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{sim_pars}{A list of simulation parameters.}

\item{replicates}{A numeric for the number of replicates for the
simulations}
}
\value{
A list output from \code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()}.
}
\description{
Runs DAISIE simulation with a novel model
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\keyword{internal}
