% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oceanic_ontogeny_di}
\alias{oceanic_ontogeny_di}
\title{Parameter space of the oceanic ontogeny diversity-independent analysis}
\format{
A data frame with 128 rows and 22 variables:
\describe{
  \item{time}{Numeric defining the length of the simulation in time units}
  \item{M}{Numeric defining the size of mainland pool, i.e. the number of
  species that can potentially colonize the island}
  \item{lac}{lambda^c (cladogenesis rate)}
  \item{mu}{mu (extinction rate)}
  \item{K}{K (carrying capacity), set K=Inf for diversity independence}
  \item{gam}{gamma (immigration rate)}
  \item{laa}{lambda^a (anagenesis rate)}
  \item{x}{the exponent for calculating extinction rate}
  \item{d}{the scaling parameter for exponent for calculating
  cladogenesis rate}
  \item{island_ontogeny}{String describing the type of island ontogeny. Can
  be "const", "beta" for a beta function describing area through time}
  \item{sea_level}{String describing the type of sea level. Can be "const" or
  "sine" for a sine function describing area through time}
  \item{max_area}{maximum area}
  \item{current_area}{current area}
  \item{peak_time}{value from 0 to 1 indicating where in the island's history
  the peak area is achieved}
  \item{total_island_age}{total island age}
  \item{sea_level_amplitude}{amplitude of area fluctuation from sea level}
  \item{sea_level_frequency}{frequency of sine wave of area change from sea
  level}
  \item{island_gradient_angle}{angle of the slope of the island}
  \item{extcutoff}{Numeric with the cutoff for the the maximum extinction
  rate preventing it from being too large and slowing down simulation}
  \item{x_s}{the probability of sampling a species from the mainland}
  \item{x_nonend}{the probability of the species sampled from the mainland
  being nonendemic}
  \item{divdepmodel}{Option divdepmodel = 'CS' runs a model with
  clade-specific carrying capacity, where diversity-dependence operates only
  within single clades, i.e. only among species originating from the same
  mainland colonist. Option divdepmodel = 'IW' runs a model with island-wide
  carrying capacity, where diversity-dependence operates within and among
  clades.}
}
}
\usage{
oceanic_ontogeny_di
}
\description{
A dataset containing the DAISIE parameters for 128 parameter sets.
}
\keyword{datasets}
