% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanic_sim.R
\name{oceanic_sim}
\alias{oceanic_sim}
\title{Runs original oceanic DAISIE simulation}
\usage{
oceanic_sim(ml, sim_pars)
}
\arguments{
\item{ml}{Output from \code{\link[DAISIE]{DAISIE_ML_CS}()}}

\item{sim_pars}{A list of simulation parameters.}
}
\value{
Output from \code{\link[DAISIE]{DAISIE_sim_constant_rate}}
}
\description{
Runs original oceanic DAISIE simulation
}
\seealso{
Other simulation functions: 
\code{\link{geodynamic_sim}()}
}
\author{
Joshua Lambert, Pedro Neves
}
\concept{simulation functions}
