% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanic_simulations.R
\name{oceanic_simulations}
\alias{oceanic_simulations}
\title{Runs original oceanic DAISIE simulation}
\usage{
oceanic_simulations(ml, simulation_pars)
}
\arguments{
\item{ml}{Output from \code{\link{DAISIE_ML_CS}}}

\item{simulation_pars}{A list of simulation parameters.}
}
\value{
Output from \code{\link{DAISIE_sim_constant_rate}}
}
\description{
Runs original oceanic DAISIE simulation
}
