% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_areas.R, R/utils.R
\name{plot_areas}
\alias{plot_areas}
\title{Plots island area function(s) through time}
\usage{
plot_areas(
  totaltime_1,
  totaltime_2 = NULL,
  area_pars_1,
  area_pars_2 = NULL,
  island_ontogeny_1 = "beta",
  island_ontogeny_2 = NULL,
  sea_level_1 = "const",
  sea_level_2 = NULL,
  resolution = 0.001,
  overlay_sea_level = FALSE
)

plot_areas(
  totaltime_1,
  totaltime_2 = NULL,
  area_pars_1,
  area_pars_2 = NULL,
  island_ontogeny_1 = "beta",
  island_ontogeny_2 = NULL,
  sea_level_1 = "const",
  sea_level_2 = NULL,
  resolution = 0.001,
  overlay_sea_level = FALSE
)
}
\arguments{
\item{totaltime_1}{Numeric with time for which island 1 would be simulated.
Usually this will correspond to a time that's smaller than the total
island age.}

\item{totaltime_2}{Time for which island 2 would be simulated. Usually this
will correspond to a time that's smaller than the total island age.
Defaults to \code{NULL}, in which case this island is not plotted.}

\item{area_pars_1}{List of area and sea-level parameters for island 1 as
created by \code{\link[DAISIE]{create_area_pars}}.}

\item{area_pars_2}{List of area and sea-level parameters for island 1 as
created by \code{\link[DAISIE]{create_area_pars}}. Defaults to \code{NULL},
in which case the second area is not plotted.}

\item{island_ontogeny_1}{A string describing the type of
island ontogeny of island 1.
Can be \code{"const"}, \code{"beta"} for a beta function describing area
through time.}

\item{island_ontogeny_2}{A string describing the type of
island ontogeny of island 2.
Can be \code{"const"}, \code{"beta"} for a beta function describing area
through time. Defaults to \code{NULL}, in which case the second island is
not plotted.}

\item{sea_level_1}{A string describing the type of sea level for island 1.
Can be \code{"const"} or \code{"sine"} for a sine function describing area
through time.}

\item{sea_level_2}{A string describing the type of sea level for island 2.
Can be \code{"const"} or \code{"sine"} for a sine function describing area
through time. Defaults to \code{NULL}, in which case the second island is
not plotted.}

\item{resolution}{A float between 0 and 1 which controls the resolution of
the island area calculated for plotting. Defaults to \code{0.001} and is
bound at \code{0.00001} to prevent unwieldy slowdowns.}

\item{overlay_sea_level}{Logical, controling if sea level curves
should be overlaid on top of corresponding ontogeny curves.}
}
\value{
a plot with the area size through time

a plot with the area size through time
}
\description{
Plots area functions through time. If only one set of area and/or sea-level
parameters are given, then only one area is plotted.

Plots area functions through time. If only one set of area and/or sea-level
parameters are given, then only one area is plotted.
}
\author{
Pedro Neves, Joshua Lambert

Pedro Neves, Joshua Lambert
}
