% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Runtime vs ED95 correlation plot}
\usage{
plot_cor(param_space_data_frame)
}
\arguments{
\item{param_space_data_frame}{A data frame with ED95 statistics of ONLY one
parameter space, together with associated runtime. Used for plotting
correlation between runtime and ED95.}
}
\value{
Saves runtime vs ED95 correlation plot to working directory as .png file.
}
\description{
Runtime vs ED95 correlation plot
}
\author{
Pedro Neves
}
