% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_log_file.R
\name{read_log_file}
\alias{read_log_file}
\title{Read log file lines}
\usage{
read_log_file(log_file_path)
}
\arguments{
\item{default_params_docs}{}
}
\value{
Character vector with the contents of the log file as returns,
  \code{\link{readLines}()}.
}
\description{
Simply sanitizes the input path and reads log files into a character vector.
}
\author{
Pedro Neves
}
