% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param_set_log.R
\name{read_param_set_log}
\alias{read_param_set_log}
\title{Reads parameter set number written in log file}
\usage{
read_param_set_log(log_lines, log_file_path)
}
\arguments{
\item{log_lines}{Character vector containing the contents of the log files
as read by \code{\link{read_log_file}()}.}

\item{log_file_path}{A string with the path for a standard Peregrine HPCC log
file.}
}
\value{
An numeric atomic vector with the parameter set number
}
\description{
Obtains the parameter set number from a standard Peregrine HPCC log file.
}
\examples{
\dontrun{
log_lines <- read_log_file("logs/robustness-20596511.log")
param_set <- read_param_set_log(
  log_lines = log_lines,
  log_file_path = "logs/robustness-20596511.log"
)
}
}
\author{
Pedro Neves
}
