% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_runtime_log.R
\name{read_runtime_log}
\alias{read_runtime_log}
\title{Calculate job elapsed time from log file}
\usage{
read_runtime_log(log_lines, log_file_path)
}
\arguments{
\item{log_lines}{Character vector containing the contents of the log files
as read by \code{\link{read_log_file}()}.}

\item{log_file_path}{A string with the path for a standard Peregrine HPCC log
file.}
}
\value{
An atomic numeric vector with the elapsed run time of the associated
  job in seconds.
}
\description{
Obtains the elapsed job time in seconds from a standard Peregrine HPCC log
file
}
\examples{
\dontrun{
log_lines <- read_log_file("logs/robustness-22353751.log")
runtime <- read_runtime_log(
  log_lines = log_lines,
  log_file_path = "logs/robustness-22353751.log"
)
}
}
\author{
Pedro Santos Neves
}
