% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the analysis with ocenanic sims and MLs}
\usage{
run_analysis(novel_sim, replicates, sim_pars)
}
\arguments{
\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{sim_pars}{A list of simulation parameters.}
}
\description{
Run the analysis with ocenanic sims and MLs
}
\seealso{
\code{\link{run_novel_sim}()}
}
\author{
Shu Xie, Joshua Lambert, Pedro Neves
}
