% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robustness_local.R
\name{run_robustness_local}
\alias{run_robustness_local}
\title{Run robustness analysis pipeline locally without saving output}
\usage{
run_robustness_local(param_space_name, param_set, replicates = 1000)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{A numeric for the number of replicates for the
simulations.}
}
\value{
A list of errors and simulation and MLE output if
\code{\link{sim_constraints}} returned TRUE or simulation
output if \code{\link{sim_constraints}} returned FALSE.
}
\description{
Run robustness analysis pipeline locally without saving output
}
\author{
Joshua Lambert, Pedro Neves
}
