% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_failed_convergence.R
\name{skip_failed_convergence}
\alias{skip_failed_convergence}
\title{Create string stating skipped step}
\usage{
skip_failed_convergence()
}
\value{
A string, with name of failed step due to failed MLE.
}
\description{
Create string stating skipped step
}
\examples{
out_string <- testit::assert(
  is.character(DAISIErobustness:::skip_failed_convergence())
)
}
\author{
Pedro Neves, Joshua Lambert
}
