% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trait_CES}
\alias{trait_CES}
\title{Parameter space of the trait-dependent analysis}
\format{
A data frame with 1664 rows and 14 variables:
\describe{
  \item{time}{Numeric defining the length of the simulation in time units}
  \item{M}{Numeric defining the size of mainland pool, i.e. the number of
  species that can potentially colonize the island}
  \item{M}{Numeric defining the size of mainland pool of species with trait
  2}
  \item{lac}{lambda^c (cladogenesis rate)}
  \item{mu}{mu (extinction rate)}
  \item{K}{K (carrying capacity), set K=Inf for diversity independence}
  \item{gam}{gamma (immigration rate)}
  \item{laa}{lambda^a (anagenesis rate)}
  \item{trans}{transition rate of trait state 1 to trait states 2}
  \item{trans2}{transition rate of trait state 2 to trait state 1}
  \item{lac2}{lambda^c (cladogenesis rate) of species with trait state 2}
  \item{mu2}{mu (extinction rate) of species with trait state 2}
  \item{gam2}{gamma (immigration rate) of species with trait state 2}
  \item{laa2}{lambda^a (anagenesis rate) of species with trait state 2}
}
}
\usage{
trait_CES
}
\description{
A dataset containing the DAISIE parameters for 1664 parameter sets.
}
\keyword{datasets}
