% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic.R
\name{all_endemic}
\alias{all_endemic}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
all_endemic(island)
}
\value{
Boolean
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
