% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic_clades.R
\name{all_endemic_clades}
\alias{all_endemic_clades}
\title{Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.}
\usage{
all_endemic_clades(sim)
}
\arguments{
\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_cr}()},
\code{\link[DAISIE]{DAISIE_sim_time_dep}()},
\code{\link[DAISIE]{DAISIE_sim_cr_shift}()},
\code{\link[DAISIE]{DAISIE_sim_trait_dep}()} or other input simulation
in the DAISIE sim format.}
}
\value{
Boolean
}
\description{
Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.
}
\author{
Joshua W. Lambert
}
