% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_recols.R
\name{any_recols}
\alias{any_recols}
\title{Determines if the island contains any recolonisations that survived to the
present}
\usage{
any_recols(island)
}
\value{
Boolean
}
\description{
Determines if the island contains any recolonisations that survived to the
present
}
