% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed95_param_set.R
\name{calc_ed95_param_set}
\alias{calc_ed95_param_set}
\title{Calculate ED95 statistic for all parameter sets of a parameter space.}
\usage{
calc_ed95_param_set(folder_path)
}
\arguments{
\item{folder_path}{A file path (use \code{file.path()} for formatting
convenience) where all the parameter sets results of a given parameter
space are stored in .RData format.}
}
\value{
A data frame with E95 statistics containing 7 columns:
\itemize{
  \item{\code{$param_space_name}}
  \item{\code{$param_set}}
  \item{\code{$stat_diff_spec_nltt}}
  \item{\code{$stat_diff_endemic_nltt}}
  \item{\code{$stat_diff_nonendemic_nltt}}
  \item{\code{$stat_diff_num_spec}}
  \item{\code{$stat_diff_num_col}}
}
}
\description{
This function will load the results of each parameter set from a given
parameter space and compute all associated ED95 statistics. All parameter
sets of the parameter space must be in the same folder, located in
\code{folder_path}.
}
\author{
Pedro Neves, Joshua W. Lambert
}
\keyword{internal}
