% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.R
\name{calc_error}
\alias{calc_error}
\title{Calcualtes error in rates, number of species-through-time, number
of endemics-through-time and non-endemics-through-time}
\usage{
calc_error(simulation_pars, ml, simulations_1, simulations_2, replicates)
}
\arguments{
\item{simulation_pars}{A list of simulation parameters.}

\item{ml}{Output from \code{\link{DAISIE_ML_CS}}}

\item{simulations_1}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{simulations_2}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{replicates}{A numeric for the number of replicates for the
simulations.}
}
\value{
A list with four error metrics
}
\description{
Calcualtes error in rates, number of species-through-time, number
of endemics-through-time and non-endemics-through-time
}
\examples{
error <- calc_error(simulation_pars = simulation_pars, ml = ml,
                      simulations_1 = simulations_1,
                      simulations_2 = simulations_2,
                      replicates = replicates)
}
