% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_metrics.R
\name{calc_error_metrics}
\alias{calc_error_metrics}
\title{Calculates metrics between error and baseline error distributions}
\usage{
calc_error_metrics(
  spec_nltt_error,
  num_spec_error,
  num_col_error,
  endemic_nltt_error,
  nonendemic_nltt_error,
  spec_baseline_nltt_error,
  num_spec_baseline_error,
  num_col_baseline_error,
  endemic_baseline_nltt_error,
  nonendemic_baseline_nltt_error
)
}
\value{
a list of numerics
}
\description{
Calculates metrics between error and baseline error distributions
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
