% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_metrics.R
\name{calc_error_metrics}
\alias{calc_error_metrics}
\title{Calculates metrics between error and baseline error distributions}
\usage{
calc_error_metrics(
  spec_error,
  endemic_error,
  nonendemic_error,
  spec_baseline_error,
  endemic_baseline_error,
  nonendemic_baseline_error
)
}
\arguments{
\item{spec_error}{A list with three elements each with a numeric
vector.}

\item{endemic_error}{A list with one element with a numeric vector.}

\item{nonendemic_error}{A list with one element with a numeric vector.}

\item{spec_baseline_error}{A list with three elements each with
a numeric vector.}

\item{endemic_baseline_error}{A list with one element with a numeric
vector.}

\item{nonendemic_baseline_error}{A list with one element with a numeric
vector.}
}
\value{
a list of numerics
}
\description{
Calculates metrics between error and baseline error distributions
}
\references{
Kolmogorov, A. (1933) Sulla determinazione empirica di una
legge di distribuzionc, 1st. Ital. Attuari. G.. 4, 1–11.

Razali and Wah (2011) "Power comparisons of shapiro-wilk,
kolmogorov-smirnov, lilliefors and anderson-darling tests"
Journal of Statistical Modeling and Analytics 2, 21-33.
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
