% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_js_dist.R
\name{calc_kl_div}
\alias{calc_kl_div}
\title{Calculates Kullback-Leibler divergence between two vectors}
\usage{
calc_kl_div(p, q)
}
\arguments{
\item{p}{Numeric vector to calculate divergence metric.}

\item{q}{Numeric vector to calculate divergence metric.}
}
\value{
numeric
}
\description{
Calculates Kullback-Leibler divergence between two vectors
}
\references{
Kullback, Solomon, and Richard A. Leibler. "On information and
  sufficiency." The annals of mathematical statistics 22.1 (1951): 79-86.
}
\author{
Pedro Neves, Joshua Lambert
}
