% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ml.R
\name{calc_ml}
\alias{calc_ml}
\title{Calculate maximum likelihood value for the rates}
\usage{
calc_ml(param_space_name, simulation_pars, simulations, verbose)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{simulation_pars}{A list of simulation parameters.}

\item{simulations}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{verbose}{logical determining whether to print
simulation statements to the console.}
}
\value{
Output from \code{\link{DAISIE::DAISIE_ML_CS}}
}
\description{
Calculate maximum likelihood value for the rates
}
\author{
Joshua Lambert, Pedro Neves
}
