% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ml.R
\name{calc_ml}
\alias{calc_ml}
\title{Calculate maximum likelihood value for the rates}
\usage{
calc_ml(sim, initial_parameters)
}
\arguments{
\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()},
\code{DAISIE_sim_trait_dependent()} or other input simulation
in the DAISIE sim format.}

\item{initial_parameters}{A numeric vector of length 5 with the initial
parameters for the MLE routine. The same as \code{initparsopt}
from \code{\link[DAISIE]{DAISIE_ML}()}.}
}
\value{
Output from \code{\link[DAISIE]{DAISIE_ML}}
}
\description{
Calculate maximum likelihood value for the rates
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
