% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error_metrics.R
\name{check_calc_error_metrics_input}
\alias{check_calc_error_metrics_input}
\title{Checks calc_error_metrics_input is correct}
\usage{
check_calc_error_metrics_input(
  spec_error,
  endemic_error,
  nonendemic_error,
  spec_baseline_error,
  endemic_baseline_error,
  nonendemic_baseline_error
)
}
\arguments{
\item{spec_error}{A list with three elements each with a numeric
vector.}

\item{endemic_error}{A list with one element with a numeric vector.}

\item{nonendemic_error}{A list with one element with a numeric vector.}

\item{spec_baseline_error}{A list with three elements each with
a numeric vector.}

\item{endemic_baseline_error}{A list with one element with a numeric
vector.}

\item{nonendemic_baseline_error}{A list with one element with a numeric
vector.}
}
\value{
Nothing. Throws error if input is not correct
}
\description{
Checks calc_error_metrics_input is correct
}
\author{
Joshua Lambert, Pedro Neves
}
