% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{check_create_folders}
\alias{check_create_folders}
\title{Checks and create results folder for output}
\usage{
check_create_folders(param_space_name, save_output)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. See
\code{\link{is_param_space_name}()} for options.}

\item{save_output}{A boolean to determine whether to save or return output.}
}
\value{
Helpful messages with info on folder status. Tries to create
  folder if needed.
}
\description{
Checks and create results folder for output
}
\seealso{
Other I/O: 
\code{\link{create_output_file_name}()},
\code{\link{save_output}()}
}
\author{
Joshua W. Lambert, Pedro Santos Neves, Shu Xie
}
\concept{I/O}
\keyword{internal}
