% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{create_output_file_name}
\alias{create_output_file_name}
\title{Generated name for saving output of \code{run_robustness()}}
\usage{
create_output_file_name(param_space_name, param_set)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. See
\code{\link{is_param_space_name}()} for options.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}
}
\value{
Character with name indicating the name of the parameter space,
  the numeric id of the param_set and whether the run passed constraints.
}
\description{
Generated name for saving output of \code{run_robustness()}
}
\examples{
testit::assert(
  DAISIErobustness:::create_output_file_name(
    param_space_name = "oceanic_ontogeny_cs",
    param_set = 1
  ) == "oceanic_ontogeny_cs_param_set_1.rds"
)
}
\seealso{
Other I/O: 
\code{\link{check_create_folders}()},
\code{\link{save_output}()}
}
\author{
Joshua Lambert, Pedro Santos Neves, Shu Xie
}
\concept{I/O}
\keyword{internal}
