% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIErobustness-package.R
\docType{package}
\name{DAISIErobustness-package}
\alias{DAISIErobustness}
\alias{DAISIErobustness-package}
\title{DAISIErobustness: Test the Robustness of DAISIE to Geodynamics and Traits}
\description{
Conducts a robustness test of the island biogeography DAISIE model
    to the effects of geodynamics or trait dependency. Tests processes such as 
    island ontogeny, sea-level change and land-bridge dynamics. 
    Contains acessory code for easy use with the University of Groningen 
    Peregrine HPCC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Neves-P/DAISIErobustness}
  \item Report bugs at \url{https://github.com/Neves-P/DAISIErobustness/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{j.w.l.lambert@rug.nl}

Authors:
\itemize{
  \item Pedro Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})
  \item Shu Xie \email{s.xie@rug.nl} (\href{https://orcid.org/0000-0001-9594-946X}{ORCID})
}

Other contributors:
\itemize{
  \item Richèl J.C. Bilderbeek \email{richel@richelbilderbeek.nl} (\href{https://orcid.org/0000-0003-1107-7049}{ORCID}) [contributor]
  \item Rampal S. Etienne \email{r.s.etienne@rug.nl} (\href{https://orcid.org/0000-0003-2142-7612}{ORCID}) [contributor]
  \item Luis Valente \email{luis.valente@naturalis.nl} (\href{https://orcid.org/0000-0003-4247-8785}{ORCID}) [contributor]
}

}
\keyword{internal}
