% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  param_space_name,
  param_set,
  replicates,
  save_output,
  output,
  param_space,
  sim_pars,
  sim,
  novel_sim,
  ml,
  ml_res_initpars_1,
  ml_res_initpars_2,
  ml_constraints,
  initial_parameters,
  sim_1,
  sim_2,
  spec_error,
  endemic_error,
  nonendemic_error,
  spec_baseline_error,
  endemic_baseline_error,
  nonendemic_baseline_error,
  distance_method,
  spec_nltt_error,
  num_spec_error,
  num_col_error,
  endemic_nltt_error,
  nonendemic_nltt_error,
  spec_baseline_nltt_error,
  num_spec_baseline_error,
  num_col_baseline_error,
  endemic_baseline_nltt_error,
  nonendemic_baseline_nltt_error,
  folder_path,
  param_set_range,
  test,
  log_file_path
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_land_bridge"},
\code{"trait_CES"}, \code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{save_output}{A boolean to determine whether to save or return output.}

\item{output}{List. Contains all the pipeline output that was able to
run up to a given point. Will be saved into .RData file by
\code{\link{save_output}()} if the \code{\link{run_robustness}()} argument
\code{save_output} is \code{TRUE}.}

\item{param_space}{tibble of the parameter space.}

\item{sim_pars}{A list of simulation parameters.}

\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()},
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()} or other input simulation
in the DAISIE sim format.}

\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}

\item{ml}{Output from \code{\link[DAISIE]{DAISIE_ML}()}.}

\item{ml_res_initpars_1}{Numeric data frame.
Results from MLE run of geodynamics with the first set of initial
parameters.}

\item{ml_res_initpars_2}{Numeric data frame.
Results from MLE run of geodynamics with the second set of initial
parameters.}

\item{ml_constraints}{Boolean. \code{TRUE} if MLE constraints pass,
\code{FALSE} otherwise.}

\item{initial_parameters}{A numeric vector of length 5 with the initial
parameters for the MLE routine. The same as \code{initparsopt}
from \code{\link[DAISIE]{DAISIE_ML}()}.}

\item{sim_1}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()}.}

\item{sim_2}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()} or
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()}.}

\item{spec_error}{A list with three elements each with a numeric
vector.}

\item{endemic_error}{A list with one element with a numeric vector.}

\item{nonendemic_error}{A list with one element with a numeric vector.}

\item{spec_baseline_error}{A list with three elements each with
a numeric vector.}

\item{endemic_baseline_error}{A list with one element with a numeric
vector.}

\item{nonendemic_baseline_error}{A list with one element with a numeric
vector.}

\item{distance_method}{From the nLTT package.
How the difference between the two nLTTs is summed:
\itemize{
 \item{\code{"abs"}: }{the absolute distance between the two nLTTs is
   summed}
 \item{\code{"squ"}: }{the squared distance between the two nLTTs is summed}
}}

\item{spec_nltt_error}{Numeric vector with the nltt error from the novel
simulation and the oceanic DAISIE simulation.}

\item{num_spec_error}{Numeric vector with the difference in the number
of species at the end of the simulation between the novel simulation and
the oceanic DAISIE simulation.}

\item{num_col_error}{Numeric vector with the difference in the number of
colonists at the end of the simulation between the novel simulation and the
 oceanic DAISIE simulation.}

\item{endemic_nltt_error}{Numeric vector with the nltt error of the
endemic species from the novel simulation and the oceanic DAISIE
simulation.}

\item{nonendemic_nltt_error}{Numeric vector with the nltt error of the
non-endemic species from the novel simulation and the oceanic DAISIE
simulation.}

\item{spec_baseline_nltt_error}{Numeric vector with the baseline nltt error
between oceanic DAISIE simulations.}

\item{num_spec_baseline_error}{Numeric vector with the difference in the
number of species at the end of the simulation between the oceanic DAISIE
simulations.}

\item{num_col_baseline_error}{Numeric vector with the difference in the
number of colonists at the end of the simulation between the oceanic DAISIE
simulations.}

\item{endemic_baseline_nltt_error}{Numeric vector with the nltt error of the
endemic species from the oceanic DAISIE simulations.}

\item{nonendemic_baseline_nltt_error}{Numeric vector with the nltt error of
the non-endemic species from the oceanic DAISIE simulations.}

\item{folder_path}{A file path (use \code{file.path()} for formatting
convenience) where all the parameter sets results of a given parameter
space are stored in .RData format.}

\item{param_set_range}{Defaults to \code{NULL}, which computes the P95
statistic for all parameter sets in the file of the parameter space located
in \code{folder_path}. Otherwise, a numeric vector of length 2, where the
first element must be lower than the second. This specifies the range of
parameter sets to load compute P95 for, i.e \code{c(1, 20)} will attempt
to compute and load parameter sets 1 to 20. If there are gaps in this
range, the function works regardless, and will return P95 for the parameter
set results within the 1 to 20 range.}

\item{test}{A boolean, defaults to `FALSE`. Set to `TRUE` for testing
purposes, to fix the seed.}

\item{log_file_path}{A string with the path for a standard Peregrine HPCC log
file.}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\keyword{internal}
