% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{load_novel_section}
\alias{load_novel_section}
\title{Load intermediate novel sim results to continue pipeline}
\usage{
load_novel_section(param_space_name, param_set)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}
}
\value{
List with output from \code{\link{run_novel_sim}()}.
}
\description{
Load intermediate novel sim results to continue pipeline
}
\seealso{
Other I/O: 
\code{\link{check_create_results_folder}()},
\code{\link{create_output_file_name}()},
\code{\link{save_output}()}
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\concept{I/O}
