% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_param_space.R
\name{load_param_space}
\alias{load_param_space}
\title{Loads the parameter space into the environment for simulations}
\usage{
load_param_space(param_space_name)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_land_bridge"},
\code{"trait_CES"}, \code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}
}
\value{
Data frame with the parameter space.
}
\description{
Loads the parameter space into the environment for simulations
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
