% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_constraints.R
\name{ml_constraints}
\alias{ml_constraints}
\title{Checks if number of valid errors meet the criteria for data sample size}
\usage{
ml_constraints(oceanic_ml)
}
\arguments{
\item{oceanic_ml}{A list of data frames with the output of DAISIE ML ran on
oceanic simulations. Elements can also be strings if the ML failed on run
or upstream.}
}
\value{
Logical \code{TRUE} if criteria are met, \code{FALSE} if not.
}
\description{
Checks if number of valid errors meet the criteria for data sample size
}
\seealso{
Other sample size constraints: 
\code{\link{sim_constraints}()}
}
\author{
Joshua Lambert, Pedro Neves
}
\concept{sample size constraints}
