% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_constraints.R
\name{ml_constraints}
\alias{ml_constraints}
\title{Checks if MLE passed or errored}
\usage{
ml_constraints(ml)
}
\arguments{
\item{ml}{Output from \code{\link[DAISIE]{DAISIE_ML}()}.}
}
\value{
Logical \code{TRUE} if criteria are met, \code{FALSE} if not.
}
\description{
Checks if MLE passed or errored
}
\author{
Joshua Lambert, Pedro Santos Neves, Shu Xie
}
\keyword{internal}
