% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_status_log.R
\name{read_status_log}
\alias{read_status_log}
\title{Calculate job elapsed time from log file}
\usage{
read_status_log(log_file_path)
}
\arguments{
\item{log_file_path}{A string with the path for a standard Peregrine HPCC log
file.}
}
\value{
A character vector with the status of the job noted in a given log
  file
}
\description{
Obtains the elapsed job time in seconds from a standard Peregrine HPCC log
file
}
\examples{
\dontrun{
status <- read_status_log(log_file_path = "logs/robustness-22353751.log")
}
}
\author{
Pedro Neves
}
