% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_status_log.R
\name{read_status_log}
\alias{read_status_log}
\title{Calculate job elapsed time from log file}
\usage{
read_status_log(log_lines)
}
\arguments{
\item{log_lines}{Character vector containing the contents of the log files
as read by \code{\link{read_log_file}()}.}
}
\value{
A character vector with the status of the job noted in a given log
  file
}
\description{
Obtains the elapsed job time in seconds from a standard Peregrine HPCC log
file
}
\author{
Pedro Neves
}
