% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the analysis with ocenanic sims and MLs}
\usage{
run_analysis(novel_sim, param_space_name, replicates, sim_pars)
}
\arguments{
\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}

\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}, \code{"trait"}.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{sim_pars}{A list of simulation parameters.}
}
\description{
Run the analysis with ocenanic sims and MLs
}
\seealso{
\code{\link{run_novel_sim}()}
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
