% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the analysis with ocenanic sims and MLs}
\usage{
run_analysis(
  novel_sim,
  param_space_name,
  replicates,
  replicate_range = NULL,
  sim_pars,
  distance_method
)
}
\arguments{
\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}

\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait_CES"},
\code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{replicate_range}{A numeric vector of length 2 or, by default,
\code{NULL} if the entire range should be used. Specifies the range of
previously saved replicates from \code{\link{run_novel_sim}()} that should
be analysed by \code{\link{run_analysis}()}.}

\item{sim_pars}{A list of simulation parameters.}

\item{distance_method}{From the nLTT package.
How the difference between the two nLTTs is summed:
\itemize{
 \item{\code{"abs"}: }{the absolute distance between the two nLTTs is
   summed}
 \item{\code{"squ"}: }{the squared distance between the two nLTTs is summed}
}}
}
\description{
Run the analysis with ocenanic sims and MLs
}
\seealso{
\code{\link{run_novel_sim}()}, \code{\link{run_robustness}()}.
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\keyword{internal}
