% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_novel_sim.R
\name{run_novel_sim}
\alias{run_novel_sim}
\title{Runs DAISIE simulation with a novel model}
\usage{
run_novel_sim(param_space_name, sim_pars)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_land_bridge"},
\code{"trait_CES"}, \code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{sim_pars}{A list of simulation parameters.}
}
\value{
A list output from \code{\link[DAISIE]{DAISIE_sim_cr}()},
\code{\link[DAISIE]{DAISIE_sim_time_dep}()},
\code{\link[DAISIE]{DAISIE_sim_cr_shift}()}, or
\code{\link[DAISIE]{DAISIE_sim_trait_dep}()}.
}
\description{
Runs DAISIE simulation with a novel model
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\keyword{internal}
