% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_oceanic_sim.R
\name{run_oceanic_sim}
\alias{run_oceanic_sim}
\title{Runs original oceanic DAISIE simulation}
\usage{
run_oceanic_sim(ml, sim_pars)
}
\arguments{
\item{ml}{Output from \code{\link[DAISIE]{DAISIE_ML}()}.}

\item{sim_pars}{A list of simulation parameters.}
}
\value{
Output from \code{\link[DAISIE]{DAISIE_sim_constant_rate}()}
}
\description{
Runs original oceanic DAISIE simulation
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
