% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robustness.R
\name{run_robustness}
\alias{run_robustness}
\title{Run robustness analysis pipeline}
\usage{
run_robustness(
  param_space_name,
  param_set,
  replicates,
  distance_method = "abs",
  save_output = TRUE
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"},
\code{"nonoceanic_land_bridge"}, \code{"trait_CES"},
\code{"trait_trans"}, or \code{"oceanic"}.
\code{"oceanic"} is used as a null simulation to test for bias in the
pipeline.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{distance_method}{From the nLTT package.
How the difference between the two nLTTs is summed:
\itemize{
 \item{\code{"abs"}: }{the absolute distance between the two nLTTs is
   summed}
 \item{\code{"squ"}: }{the squared distance between the two nLTTs is summed}
}}

\item{save_output}{A boolean to determine whether to save or return output.}
}
\value{
A list of errors and simulation and MLE output.
}
\description{
Run robustness analysis pipeline
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
