% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_robustness.R
\name{run_robustness}
\alias{run_robustness}
\title{Run robustness analysis pipeline}
\usage{
run_robustness(
  param_space_name,
  param_set,
  replicates,
  pipeline = "full",
  novel_sim = NULL,
  save_output = TRUE
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}, \code{"trait"}.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}

\item{replicates}{A numeric for the number of replicates for the
simulations}

\item{pipeline}{A string stating what sections of the pipeline should
run. Default is \code{"full"}, meaning the entire pipeline, including the
novel model and analysis will run. \code{"novel_sim"} Indicates only the
novel simulation section should run, calling \code{\link{run_novel_sim}()}.
\code{"analysis"} runs just the estimation, oceanic simulations and error
calculation sections of the pipeline calling \code{\link{run_analysis}()}.}

\item{novel_sim}{A list of simulation output in the DAISIE simulation format
for which the robustness against standard oceanic DAISIE is to be
determined.
Can be generated from a preset of models available in
\code{\link{run_novel_sim}()}, or loaded from a file, if in the correct
format.}

\item{save_output}{A boolean to determine whether to save or return output.}
}
\value{
A list of errors and simulation and MLE output if
\code{\link{sim_constraints}()} and \code{\link{ml_constraints}()} returned
TRUE or simulation output if \code{\link{sim_constraints}()} returned FALSE.
}
\description{
Run robustness analysis pipeline
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
