% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output.R
\name{save_output}
\alias{save_output}
\title{Save \code{run_robustness()} output}
\usage{
save_output(
  output_file,
  param_space_name,
  param_set,
  sim_constraints,
  ml_constraints
)
}
\arguments{
\item{param_space_name}{A string with the parameter space to run. Can
be \code{"oceanic_ontogeny"}, \code{"oceanic_sea_level"},
\code{"oceanic_ontogeny_sea_level"},
\code{"nonoceanic"}, \code{"nonoceanic_sea_level"}, or
\code{"nonoceanic_land_bridge"}.}

\item{param_set}{A numeric with the line corresponding to parameter set to
run, as found in the file named in \code{param_space}.}
}
\value{
.RData file named by \code{\link{create_output_filename}()} in
default location. See details for more information on filesystem. This
}
\description{
Save \code{run_robustness()} output
}
\details{
Files are saved following the standardized file structure of the
TECE lab, as implemented in \code{\link[jap]{folder_structure}()}.
}
\seealso{
Other I/O: 
\code{\link{check_create_results_folder}()},
\code{\link{create_output_filename}()}
}
\author{
Pedro Neves, Joshua Lambert
}
\concept{I/O}
