% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_constraints.R
\name{sim_constraints}
\alias{sim_constraints}
\title{Checks if the simulations meet the criteria for data sample size}
\usage{
sim_constraints(sim, replicates)
}
\arguments{
\item{sim}{A list of simulation output from
\code{\link[DAISIE]{DAISIE_sim_constant_rate}()},
\code{\link[DAISIE]{DAISIE_sim_time_dependent}()},
\code{\link[DAISIE]{DAISIE_sim_constant_rate_shift}()},
\code{\link[DAISIE]{DAISIE_sim_trait_dependent}()} or other input simulation
in the DAISIE sim format.}

\item{replicates}{A numeric for the number of replicates for the
simulations}
}
\value{
Logical TRUE if criteria are met, FALSE if not.
}
\description{
Checks if the simulations meet the criteria for data sample size
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
