% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_constraints.R
\name{simulation_constraints}
\alias{simulation_constraints}
\title{Checks if the simulations meet the criteria for data sample size}
\usage{
simulation_constraints(simulations, replicates)
}
\arguments{
\item{simulations}{A list of simulation output from
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}} or
\code{\link{DAISIE_sim_constant_rate_shift}}}

\item{replicates}{A numeric for the number of replicates for the
simulations.}
}
\value{
Boolean
}
\description{
Checks if the simulations meet the criteria for data sample size
}
